+++ sdk/identity/azure-identity/azure/identity/_internal/__init__.py
@@ -2,6 +2,16 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
 # ------------------------------------
+import os
+
+from .._constants import EnvironmentVariables, KnownAuthorities
+
+
+def get_default_authority():
+    return os.environ.get(EnvironmentVariables.AZURE_AUTHORITY_HOST, KnownAuthorities.AZURE_PUBLIC_CLOUD)
+
+
+# pylint:disable=wrong-import-position
 from .aad_client import AadClient
 from .aad_client_base import AadClientBase
 from .auth_code_redirect_handler import AuthCodeRedirectServer
@@ -14,7 +24,7 @@ def _scopes_to_resource(*scopes):
     """Convert an AADv2 scope to an AADv1 resource"""
 
     if len(scopes) != 1:
-        raise ValueError("This credential supports only one scope per token request")
+        raise ValueError("This credential requires exactly one scope per token request.")
 
     resource = scopes[0]
     if resource.endswith("/.default"):