+++ sdk/keyvault/azure-mgmt-keyvault/tests/test_cli_mgmt_keyvault.py
@@ -28,26 +28,31 @@ class MgmtKeyVaultTest(AzureMgmtTestCase):
     def setUp(self):
         super(MgmtKeyVaultTest, self).setUp()
         self.mgmt_client = self.create_mgmt_client(
-            azure.mgmt.keyvault.KeyVault
+            azure.mgmt.keyvault.KeyVaultManagementClient
         )
     
     @ResourceGroupPreparer(location=AZURE_LOCATION)
     def test_keyvault(self, resource_group):
 
-        SERVICE_NAME = "myapimrndxyz"
+        SUBSCRIPTION_ID = self.settings.SUBSCRIPTION_ID
+        RESOURCE_GROUP = resource_group.name
+        VAULT_NAME = "myVaultSss"
+        OPERATION_KIND = "add"
+        LOCATION = "eastus"
+        PRIVATE_ENDPOINT_CONNECTION_NAME = "myPrivateEndpointConnection"
 
-        # Create a new vault or update an existing vault[put]
+        # /Vaults/put/Create a new vault or update an existing vault[put]
         BODY = {
           "location": "westus",
           "properties": {
-            "tenant_id": "00000000-0000-0000-0000-000000000000",
+            "tenant_id": "72f988bf-86f1-41af-91ab-2d7cd011db47",
             "sku": {
               "family": "A",
               "name": "standard"
             },
             "access_policies": [
               {
-                "tenant_id": "00000000-0000-0000-0000-000000000000",
+                "tenant_id": "72f988bf-86f1-41af-91ab-2d7cd011db47",
                 "object_id": "00000000-0000-0000-0000-000000000000",
                 "permissions": {
                   "keys": [
@@ -102,10 +107,10 @@ class MgmtKeyVaultTest(AzureMgmtTestCase):
             "enabled_for_template_deployment": True
           }
         }
-        result = self.mgmt_client.vaults.create_or_update(resource_group.name, VAULT_NAME, BODY)
+        result = self.mgmt_client.vaults.create_or_update(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME, parameters=BODY)
         result = result.result()
 
-        # Create or update a vault with network acls[put]
+        # /Vaults/put/Create or update a vault with network acls[put]
         BODY = {
           "location": "westus",
           "properties": {
@@ -127,7 +132,7 @@ class MgmtKeyVaultTest(AzureMgmtTestCase):
               ],
               "virtual_network_rules": [
                 {
-                  "id": "/subscriptions/" + SUBSCRIPTION_ID + "/resourceGroups/" + RESOURCE_GROUP + "/providers/Microsoft.Network/virtualNetworks/" + VIRTUAL_NETWORK_NAME + "/subnets/" + SUBNET_NAME + ""
+                  "id": "/subscriptions/" + SUBSCRIPTION_ID + "/resourceGroups/" + RESOURCE_GROUP + "/providers/Microsoft.Network/virtualNetworkstest-vnet/subnetssubnet1"
                 }
               ]
             },
@@ -136,13 +141,11 @@ class MgmtKeyVaultTest(AzureMgmtTestCase):
             "enabled_for_template_deployment": True
           }
         }
-        result = self.mgmt_client.vaults.create_or_update(resource_group.name, VAULT_NAME, BODY)
-        result = result.result()
+        #result = self.mgmt_client.vaults.create_or_update(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME, parameters=BODY)
+        #result = result.result()
 
-        # Add an access policy, or update an access policy with new permissions[put]
-        BODY = {
-          "properties": {
-            "access_policies": [
+        # /Vaults/put/Add an access policy, or update an access policy with new permissions[put]
+        PROPERTIES = [
               {
                 "tenant_id": "00000000-0000-0000-0000-000000000000",
                 "object_id": "00000000-0000-0000-0000-000000000000",
@@ -158,134 +161,123 @@ class MgmtKeyVaultTest(AzureMgmtTestCase):
                   ]
                 }
               }
-            ]
-          }
-        }
-        result = self.mgmt_client.vaults.update_access_policy(resource_group.name, VAULT_NAME, ACCESS_POLICY_NAME, BODY)
+        ]
+        # result = self.mgmt_client.vaults.update_access_policy(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME, operation_kind=OPERATION_KIND, properties=PROPERTIES)
 
-        # KeyVaultPutPrivateEndpointConnection[put]
+        # /PrivateEndpointConnections/put/KeyVaultPutPrivateEndpointConnection[put]
         BODY = {
-          "properties": {
-            "private_link_service_connection_state": {
-              "status": "Approved",
-              "description": "My name is Joe and I'm approving this."
-            }
+          "private_link_service_connection_state": {
+            "status": "Approved",
+            "description": "My name is Joe and I'm approving this."
           }
         }
-        result = self.mgmt_client.private_endpoint_connections.put(resource_group.name, VAULT_NAME, PRIVATE_ENDPOINT_CONNECTION_NAME, BODY)
+        # result = self.mgmt_client.private_endpoint_connections.put(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME, private_endpoint_connection_name=PRIVATE_ENDPOINT_CONNECTION_NAME, properties=BODY)
 
-        # KeyVaultGetPrivateEndpointConnection[get]
-        result = self.mgmt_client.private_endpoint_connections.get(resource_group.name, VAULT_NAME, PRIVATE_ENDPOINT_CONNECTION_NAME)
+        # /PrivateEndpointConnections/get/KeyVaultGetPrivateEndpointConnection[get]
+        # result = self.mgmt_client.private_endpoint_connections.get(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME, private_endpoint_connection_name=PRIVATE_ENDPOINT_CONNECTION_NAME)
 
-        # KeyVaultListPrivateLinkResources[get]
-        result = self.mgmt_client.private_link_resources.list_by_vault(resource_group.name, VAULT_NAME)
+        # /PrivateLinkResources/get/KeyVaultListPrivateLinkResources[get]
+        # result = self.mgmt_client.private_link_resources.list_by_vault(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME)
 
-        # Retrieve a deleted vault[get]
-        result = self.mgmt_client.vaults.get_deleted(LOCATION_NAME, DELETED_VAULT_NAME)
+        # /Vaults/get/Retrieve a vault[get]
+        result = self.mgmt_client.vaults.get(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME)
 
-        # Retrieve a vault[get]
-        result = self.mgmt_client.vaults.get(resource_group.name, VAULT_NAME)
+        # /Vaults/get/List vaults in the specified resource group[get]
+        result = self.mgmt_client.vaults.list_by_resource_group(resource_group_name=RESOURCE_GROUP, top="1")
 
-        # List vaults in the specified resource group[get]
-        result = self.mgmt_client.vaults.list_by_resource_group(resource_group.name)
-
-        # List deleted vaults in the specified subscription[get]
+        # /Vaults/get/List deleted vaults in the specified subscription[get]
         result = self.mgmt_client.vaults.list_deleted()
 
-        # List vaults in the specified subscription[get]
-        result = self.mgmt_client.vaults.list_by_subscription()
+        # /Vaults/get/List vaults in the specified subscription[get]
+        result = self.mgmt_client.vaults.list_by_subscription(top="1")
 
-        # List vaults in the specified subscription[get]
-        result = self.mgmt_client.vaults.list_by_subscription()
+        # /Vaults/get/List vaults in the specified subscription[get]
+        result = self.mgmt_client.vaults.list_by_subscription(top="1")
 
-        # Lists available Rest API operations.[get]
+        # /Operations/get/Lists available Rest API operations.[get]
         result = self.mgmt_client.operations.list()
 
-        # Purge a deleted vault[post]
-        result = self.mgmt_client.vaults.purge_deleted(LOCATION_NAME, DELETED_VAULT_NAME)
-        result = result.result()
-
-        # Update an existing vault[patch]
-        BODY = {
-          "properties": {
-            "tenant_id": "00000000-0000-0000-0000-000000000000",
-            "sku": {
-              "family": "A",
-              "name": "standard"
-            },
-            "access_policies": [
-              {
-                "tenant_id": "00000000-0000-0000-0000-000000000000",
-                "object_id": "00000000-0000-0000-0000-000000000000",
-                "permissions": {
-                  "keys": [
-                    "encrypt",
-                    "decrypt",
-                    "wrapKey",
-                    "unwrapKey",
-                    "sign",
-                    "verify",
-                    "get",
-                    "list",
-                    "create",
-                    "update",
-                    "import",
-                    "delete",
-                    "backup",
-                    "restore",
-                    "recover",
-                    "purge"
-                  ],
-                  "secrets": [
-                    "get",
-                    "list",
-                    "set",
-                    "delete",
-                    "backup",
-                    "restore",
-                    "recover",
-                    "purge"
-                  ],
-                  "certificates": [
-                    "get",
-                    "list",
-                    "delete",
-                    "create",
-                    "import",
-                    "update",
-                    "managecontacts",
-                    "getissuers",
-                    "listissuers",
-                    "setissuers",
-                    "deleteissuers",
-                    "manageissuers",
-                    "recover",
-                    "purge"
-                  ]
-                }
+        # /Vaults/patch/Update an existing vault[patch]
+        PROPERTIES = {
+          "tenant_id": "72f988bf-86f1-41af-91ab-2d7cd011db47",
+          "sku": {
+            "family": "A",
+            "name": "standard"
+          },
+          "access_policies": [
+            {
+              "tenant_id": "00000000-0000-0000-0000-000000000000",
+              "object_id": "00000000-0000-0000-0000-000000000000",
+              "permissions": {
+                "keys": [
+                  "encrypt",
+                  "decrypt",
+                  "wrapKey",
+                  "unwrapKey",
+                  "sign",
+                  "verify",
+                  "get",
+                  "list",
+                  "create",
+                  "update",
+                  "import",
+                  "delete",
+                  "backup",
+                  "restore",
+                  "recover",
+                  "purge"
+                ],
+                "secrets": [
+                  "get",
+                  "list",
+                  "set",
+                  "delete",
+                  "backup",
+                  "restore",
+                  "recover",
+                  "purge"
+                ],
+                "certificates": [
+                  "get",
+                  "list",
+                  "delete",
+                  "create",
+                  "import",
+                  "update",
+                  "managecontacts",
+                  "getissuers",
+                  "listissuers",
+                  "setissuers",
+                  "deleteissuers",
+                  "manageissuers",
+                  "recover",
+                  "purge"
+                ]
               }
-            ],
-            "enabled_for_deployment": True,
-            "enabled_for_disk_encryption": True,
-            "enabled_for_template_deployment": True
-          }
+            }
+          ],
+          "enabled_for_deployment": True,
+          "enabled_for_disk_encryption": True,
+          "enabled_for_template_deployment": True
         }
-        result = self.mgmt_client.vaults.update(resource_group.name, VAULT_NAME, BODY)
+        # result = self.mgmt_client.vaults.update(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME, properties=PROPERTIES)
 
-        # Validate a vault name[post]
-        BODY = {
-          "name": "sample-vault",
-          "type": "Microsoft.KeyVault/vaults"
-        }
-        result = self.mgmt_client.vaults.check_name_availability(BODY)
+        # /Vaults/post/Validate a vault name[post]
+        result = self.mgmt_client.vaults.check_name_availability(name="sample-vault", type="Microsoft.KeyVault/vaults")
 
-        # KeyVaultDeletePrivateEndpointConnection[delete]
-        result = self.mgmt_client.private_endpoint_connections.delete(resource_group.name, VAULT_NAME, PRIVATE_ENDPOINT_CONNECTION_NAME)
-        result = result.result()
+        # /PrivateEndpointConnections/delete/KeyVaultDeletePrivateEndpointConnection[delete]
+        # result = self.mgmt_client.private_endpoint_connections.delete(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME, private_endpoint_connection_name=PRIVATE_ENDPOINT_CONNECTION_NAME)
+        # result = result.result()
 
-        # Delete a vault[delete]
-        result = self.mgmt_client.vaults.delete(resource_group.name, VAULT_NAME)
+        # /Vaults/delete/Delete a vault[delete]
+        result = self.mgmt_client.vaults.delete(resource_group_name=RESOURCE_GROUP, vault_name=VAULT_NAME)
 
+        # /Vaults/get/Retrieve a deleted vault[get]
+        # result = self.mgmt_client.vaults.get_deleted(location=LOCATION, vault_name=VAULT_NAME)
+
+        # /Vaults/post/Purge a deleted vault[post]
+        # result = self.mgmt_client.vaults.purge_deleted(location=LOCATION, vault_name=VAULT_NAME)
+        # result = result.result()
 
 #------------------------------------------------------------------------------
 if __name__ == '__main__':