+++ sdk/search/azure-search-documents/azure/search/documents/_index/aio/_search_index_client_async.py
@@ -7,22 +7,22 @@ from typing import cast, List, TYPE_CHECKING
 
 import six
 
-from azure.core.pipeline.policies import HeadersPolicy
 from azure.core.tracing.decorator_async import distributed_trace_async
 from ._paging import AsyncSearchItemPaged, AsyncSearchPageIterator
 from .._generated.aio import SearchIndexClient as _SearchIndexClient
 from .._generated.models import IndexBatch, IndexingResult, SearchRequest
 from .._index_documents_batch import IndexDocumentsBatch
 from .._queries import AutocompleteQuery, SearchQuery, SuggestQuery
-from ..._version import VERSION
+from ..._headers_mixin import HeadersMixin
+from ..._version import SDK_MONIKER
 
 if TYPE_CHECKING:
     # pylint:disable=unused-import,ungrouped-imports
     from typing import Any, Union
-    from .._credential import SearchApiKeyCredential
+    from azure.core.credentials import AzureKeyCredential
 
 
-class SearchIndexClient(object):
+class SearchIndexClient(HeadersMixin):
     """A client to interact with an existing Azure search index.
 
     :param endpoint: The URL endpoint of an Azure search service
@@ -30,7 +30,7 @@ class SearchIndexClient(object):
     :param index_name: The name of the index to connect to
     :type index_name: str
     :param credential: A credential to authorize search client requests
-    :type credential: SearchApiKeyCredential
+    :type credential: ~azure.core.credentials.AzureKeyCredential
 
     .. admonition:: Example:
 
@@ -42,24 +42,16 @@ class SearchIndexClient(object):
             :caption: Creating the SearchIndexClient with an API key.
     """
 
-    def __init__(self, endpoint, index_name, credential, **kwargs):
-        # type: (str, str, SearchApiKeyCredential, **Any) -> None
+    _ODATA_ACCEPT = "application/json;odata.metadata=none"  # type: str
 
-        headers_policy = HeadersPolicy(
-            {
-                "api-key": credential.api_key,
-                "Accept": "application/json;odata.metadata=none",
-            }
-        )
+    def __init__(self, endpoint, index_name, credential, **kwargs):
+        # type: (str, str, AzureKeyCredential, **Any) -> None
 
         self._endpoint = endpoint  # type: str
         self._index_name = index_name  # type: str
+        self._credential = credential  # type: AzureKeyCredential
         self._client = _SearchIndexClient(
-            endpoint=endpoint,
-            index_name=index_name,
-            headers_policy=headers_policy,
-            sdk_moniker="search/{}".format(VERSION),
-            **kwargs
+            endpoint=endpoint, index_name=index_name, sdk_moniker=SDK_MONIKER, **kwargs
         )  # type: _SearchIndexClient
 
     def __repr__(self):
@@ -82,6 +74,7 @@ class SearchIndexClient(object):
 
         :rtype: int
         """
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         return int(await self._client.documents.count(**kwargs))
 
     @distributed_trace_async
@@ -104,6 +97,7 @@ class SearchIndexClient(object):
                 :dedent: 4
                 :caption: Get a specific document from the search index.
         """
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         result = await self._client.documents.get(
             key=key, selected_fields=selected_fields, **kwargs
         )
@@ -153,7 +147,7 @@ class SearchIndexClient(object):
                     repr(query)
                 )
             )
-
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         return AsyncSearchItemPaged(
             self._client, query, kwargs, page_iterator_class=AsyncSearchPageIterator
         )
@@ -181,6 +175,7 @@ class SearchIndexClient(object):
                 "Expected a SuggestQuery for 'query', but got {}".format(repr(query))
             )
 
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         response = await self._client.documents.suggest_post(
             suggest_request=query.request, **kwargs
         )
@@ -212,6 +207,7 @@ class SearchIndexClient(object):
                 )
             )
 
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         response = await self._client.documents.autocomplete_post(
             autocomplete_request=query.request, **kwargs
         )
@@ -241,6 +237,8 @@ class SearchIndexClient(object):
         """
         batch = IndexDocumentsBatch()
         batch.add_upload_documents(documents)
+
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         results = await self.index_documents(batch, **kwargs)
         return cast(List[IndexingResult], results)
 
@@ -272,6 +270,8 @@ class SearchIndexClient(object):
         """
         batch = IndexDocumentsBatch()
         batch.add_delete_documents(documents)
+
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         results = await self.index_documents(batch, **kwargs)
         return cast(List[IndexingResult], results)
 
@@ -299,6 +299,8 @@ class SearchIndexClient(object):
         """
         batch = IndexDocumentsBatch()
         batch.add_merge_documents(documents)
+
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         results = await self.index_documents(batch, **kwargs)
         return cast(List[IndexingResult], results)
 
@@ -317,6 +319,8 @@ class SearchIndexClient(object):
         """
         batch = IndexDocumentsBatch()
         batch.add_merge_or_upload_documents(documents)
+
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         results = await self.index_documents(batch, **kwargs)
         return cast(List[IndexingResult], results)
 
@@ -330,6 +334,8 @@ class SearchIndexClient(object):
         :rtype:  List[IndexingResult]
         """
         index_documents = IndexBatch(actions=batch.actions)
+
+        kwargs["headers"] = self._merge_client_headers(kwargs.get("headers"))
         batch_response = await self._client.documents.index(
             batch=index_documents, **kwargs
         )