+++ sdk/storage/azure-storage-blob/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 12.3.1 (Unreleased)
+
+**Fixes**
+- Fixed issue where batch requests could not be combined with token credentials (#9534)
+
+
 ## 12.3.0 (2020-03-10)
 
 **New features**